arm, arm926ejs: make thumb mode compileable
in thumb mode compiler says for example for arch/arm/lib/cache-cp15.c
when enabling CONFIG_SYS_THUMB_BUILD:

{standard input}: Assembler messages:
{standard input}:373: Error: selected processor does not support Thumb mode `mrc p15,0,r4,c1,c0,0'
{standard input}:416: Error: selected processor does not support Thumb mode `mcr p15,0,r3,c2,c0,0'

so, if caches are disabled, do not use this command on arm926ejs.
used on at91 in SPL, to reduce size of SPL.

Signed-off-by: Heiko Schocher <hs@denx.de>
Patch-cc: Albert Aribaud <albert.u.boot@aribaud.net>
Patch-cc: Tom Rini <trini@ti.com>

Series-to: u-boot
Series-version: 2
Cover-letter:
arm, at91: taurus board updates

This series enables THUMB mode in SPL for the taurus board. On this
hw are 16k only for SPL code, so THUMB mode is needed, because we
need to detect the nor flash on the spi bus. This is needed because
we want to erase sector 0 of the spi nor flash if the recovery button
is pressed in SPL.
END

Series-changes: 2
rebased against 2a82ec77d27ef5: "Prepare v2015.01-rc2"