imx6: clock: add CAN clock setting
The CAN clock is derived from PLL3_60M (60 MHz, derived from PLL3
with fixed divider of 8) with an adjustable divider. By default,
the value of the divider is 2. So the default CAN clock is 30 Mhz.

It is implemented to change the CAN clock from Kconfig if requested
and the divider value is parameterized.

Signed-off-by: Leo Ruan <tingquan.ruan@cn.bosch.com>
adapted to 201611 from hs
Signed-off-by: Heiko Schocher <hs@denx.de>